TST: Add testing of numba-path code
Adds pure python paths so numba-compiled code can be directly tested

+++ arch/bootstrap/_samplers.pyx
@@ -5,7 +5,7 @@ cimport cython
 @cython.boundscheck(False)
 @cython.wraparound(False)
 @cython.cdivision(True)
-def stationary_bootstrap_sample(int[:] indices,
+def stationary_bootstrap_sample(long[:] indices,
                                 double[:] u,
                                 double p):
     """
